Verplaats zijbalk meescrollen naar entry
